+++ packages/react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [6.101.0](https://github.com/factorialco/f0/compare/f0-react-v6.100.1...f0-react-v6.101.0) (2026-09-14)
+
+
+### Features
+
+* **F0Select:** add multi-level grouping and trigger label control ([#5530](https://github.com/factorialco/f0/issues/5530)) ([5276de8](https://github.com/factorialco/f0/commit/5276de87b7be56e2dce4551d8f3c1ced6bf58c06))
+
+## [6.100.1](https://github.com/factorialco/f0/compare/f0-react-v6.100.0...f0-react-v6.100.1) (2026-09-14)
+
+
+### Bug Fixes
+
+* **virtuoso:** Revert "fix(virtuoso): revert "chore(deps): drop the react-virtuoso patch for 4.18.13"" ([#5526](https://github.com/factorialco/f0/issues/5526)) ([5c763a2](https://github.com/factorialco/f0/commit/5c763a2ea386d9cd742a2fe2c3e6f979ce0f796d))
+
 ## [6.100.0](https://github.com/factorialco/f0/compare/f0-react-v6.99.0...f0-react-v6.100.0) (2026-09-14)